Expected a warning
Hg-commit: 4b159146106cd3fbe47a07f062c7432ffdeb0e27

+++ src/sys/viewer/impls/socket/send.c
@@ -384,7 +384,7 @@ PetscErrorCode PETSCSYS_DLLEXPORT PetscViewerSocketSetConnection(PetscViewer v,c
     if (tflg) {
       PetscInt pport;
       ierr = PetscOptionsAtoi(portn,&pport);CHKERRQ(ierr);
-      port = pport;
+      port = (int)pport;
     } else {
       port = PETSCSOCKETDEFAULTPORT;
     }
@@ -584,7 +584,7 @@ PetscErrorCode PETSCSYS_DLLEXPORT PetscWebServeRequest(int port)
     PetscFunctionReturn(0);
   }   
 
-  fseek(fd, 0, SEEK_CUR); // Force change of stream direction
+  fseek(fd, 0, SEEK_CUR); /* Force change of stream direction */
 
   ierr = PetscStrcmp(method,"GET",&flg);
   if (!flg) {
